+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next release
+
+ - Add junk detection for AndOrReplacement mutator #509
+
 ## [0.2.0] - 02 May 2019
 
  - Add FreeBSD Package #503